feat(睡眠功能): 新增定时唤醒时间支持并完善休眠控制
1. 扩展sleep_time字段从uint8_t改为uint16_t，支持10分钟到48小时范围
2. 新增hal层定时器唤醒配置接口，实现基于分钟的唤醒时间设置
3. 完善自动休眠逻辑，根据配置启用定时器唤醒
4. 补充BLE休眠相关控制命令的处理和查询响应，包括模式、开关和时间配置

+++ firmware/frame_film/components/film_service/inc/service_param.h
@@ -33,7 +33,7 @@ typedef struct
 {
     uint8_t sleep_mode;        // 休眠模式开关 0：关闭 1：开启
     uint8_t sleep_auto;        // 定时唤醒开关 0：关闭 1：开启
-    uint8_t sleep_time;        // 定时唤醒时间（单位分钟）
+    uint16_t sleep_time;       // 定时唤醒时间（单位分钟 10min - 48h(48*60min)）
 } ServiceSleep_Def_t;
 
 #pragma pack(4)
